app/src/main/java/com/example/SimpleVoiceRecorder/gotov/Model/Sound.java
package com.example.SimpleVoiceRecorder.gotov.Model;

import androidx.room.Entity;
import androidx.room.PrimaryKey;

@Entity
public class Sound {
    @PrimaryKey(autoGenerate = true)
    public long id;
    private String recordName;
    private String filePath;
    private String length;
    private String timeAdded;

    public long getId() {
        return id;
    }

    public String getRecordName() {
        return recordName;
    }

    public void setRecordName(String recordName) {
        this.recordName = recordName;
    }

    public String getFilePath() {
        return filePath;
    }

    public void setFilePath(String filePath) {
        this.filePath = filePath;
    }

    public String getLength() {
        return length;
    }

    public void setLength(String length) {
        this.length = length;
    }

    public String getTimeAdded() {
        return timeAdded;
    }

    public void setTimeAdded(String timeAdded) {
        this.timeAdded = timeAdded;
    }
}
